// frontend/src/api-v1/models/teamDto.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Rebus backend
 * OpenAPI spec version: 2.2.0
 */

/**
 * The data that the server will send in the response, can be null
 */
export interface TeamDto {
  /** ID of the team */
  id: number;
  /** Name of the team */
  name: string;
}