// frontend/src/api-v2/models/answerDto.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * rebus-backend
 * Backend for the Rebus application, in Node.js
 * OpenAPI spec version: 2.3.0
 */

/**
 * Answer DTO.
 */
export interface AnswerDto {
  /** @nullable */
  answer: string | null;
  challengeId: number;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  score: number | null;
}