// frontend/src/api-v2/models/teamAnswerDto.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * rebus-backend
 * Backend for the Rebus application, in Node.js
 * OpenAPI spec version: 2.3.0
 */
import type { AnswerDto } from "./answerDto";

/**
 * Team answer DTO.
 */
export interface TeamAnswerDto {
  answers: AnswerDto[];
  teamId: number;
}